Created folders to organize components files
