+++ frontend/src/components/models/newsFeedList/NewsFeedListButtons/index.ts
@@ -1,0 +1,1 @@
+export { SearchFormCloseButton, SearchFormOpenButton } from './NewsFeedListButtons';
