+++ frontend/src/components/models/newsFeedList/NewsFeedListSubs/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NewsFeedListSubs';
