+++ frontend/src/components/models/newsFeedList/NewsFeedsItem/index.ts
