+++ frontend/src/components/pages/Lp/index.ts
