+++ frontend/src/components/pages/NewsContent/index.ts
