+++ frontend/src/components/pages/Search/index.ts
