+++ frontend/src/components/ui/Footer/Footer.tsx
