+++ frontend/src/components/ui/Footer/index.ts
