+++ frontend/src/components/ui/LoadingIndicator/LoadingIndicator.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import '../../styles/loadingIndicator.scss';
+import '../../../styles/loadingIndicator.scss';
 
 const LoadingIndicator: FC<{ isOverlay: boolean; }> = ({ isOverlay }) => {
