+++ frontend/src/main.tsx
@@ -4,13 +4,13 @@ import { store } from './redux/app/configureStore';
 import { Provider } from 'react-redux';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
-import './index.scss';
+import SectionLayout from './components/ui/SectionLayout';
+import Lp from './components/pages/Lp';
+import Search from './components/pages/Search';
+import NewsFeedsList from './components/pages/NewsFeedsList';
+import NewsContent from './components/pages/NewsContent';
 
-import SectionLayout from './components/SectionLayout';
-import Lp from './pages/Lp';
-import Search from './pages/Search';
-import NewsFeedsList from './pages/NewsFeedsList';
-import NewsContent from './pages/NewsContent';
+import './index.scss';
 
 const router = createBrowserRouter([
   {
